pyproject.toml
[project]
name = "terraform-sample"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bandit>=1.8.6",
    "fastapi[standard]>=0.116.1",
    "pytest-bdd>=8.1.0",
    "pytest-playwright>=0.7.1",
]
[tool.pytest.ini_options]
testpaths = [
    "tests",
    "features"
]
